refactor ResponseTimeDashboard (move logic to dedicated file)
